Organize methods between parent and child classes
